Added more validation in terms of disabled campaigns

+++ mailprime/mailer/recipient_views.py
@@ -25,23 +25,30 @@ def user_campaign_recipients(request, param_username, param_campaign_pk):
 
 		recipients = Recipient.objects.filter(campaign = campaign).order_by('email')
 
-		# Looking for the 'delete' GET parameter, so we can remove the recipient
-		recipient_to_remove = request.GET.get('delete')
-		print recipient_to_remove
+		# Checking to see if the campaign is disabled
+		if campaign.active:
 
-		# Check to make sure, recip_pk belongs to campaign.
-		# If so, delete it from the database.
-		if recipient_to_remove is not None:
+			# Looking for the 'delete' GET parameter, so we can remove the recipient
+			recipient_to_remove = request.GET.get('delete')
+			print recipient_to_remove
 
-			for x in recipients:
+			# Check to make sure, recip_pk belongs to campaign.
+			# If so, delete it from the database.
+			if recipient_to_remove is not None:
 
-				if int(recipient_to_remove) == x.pk:
+				for x in recipients:
 
-					messages.add_message(request, messages.SUCCESS, 'Successfully Removed {0}'.format( x.email ))
-					x.delete()
+					if int(recipient_to_remove) == x.pk:
 
-					# Update recipients queryset
-					recipients = Recipient.objects.filter(campaign = campaign).order_by('email')
+						messages.add_message(request, messages.SUCCESS, 'Successfully Removed {0}'.format( x.email ))
+						x.delete()
+
+						# Update recipients queryset
+						recipients = Recipient.objects.filter(campaign = campaign).order_by('email')
+		else:
+
+			# Generate warning because campaign is disabled
+			messages.add_message(request, messages.SUCCESS, 'Campaign disabled by MailPrime Administrator')
 
 		# Page varaibles and render page
 		page_vars['campaign'] = campaign
@@ -58,14 +65,14 @@ def upload_recipients(request, param_username, param_campaign_pk):
 
 	if current_user(request) and request.user.username == param_username:
 
+		# Capturing Campaign that will be used with the recipient entries
+		try:
+			campaign = Campaign.objects.get(pk=param_campaign_pk, user=request.user)
+		except Campaign.DoesNotExist:
+			raise Http404
+
 		# User is requesting page, render form
 		if request.method == 'GET':
-
-			# Capturing Campaign that will be used with the recipient entries
-			try:
-				campaign = Campaign.objects.get(pk=param_campaign_pk, user=request.user)
-			except Campaign.DoesNotExist:
-				raise Http404
 
 			page_vars['campaign'] = campaign
 			page_vars['form'] = ContactUploadForm()
@@ -75,6 +82,15 @@ def upload_recipients(request, param_username, param_campaign_pk):
 
 		# User is uploading from form, parse data
 		elif request.method == 'POST':
+
+			# Check to see if campaign is disabled
+			if campaign.active == False:
+
+				#Generate warning, redirect.
+				messages.add_message(request, messages.SUCCESS, 'Campaign disabled by MailPrime Administrator')
+				return HttpResponseRedirect('/{0}/campaign-{1}/recipients/upload'.format(request.user.username, campaign.pk))
+
+			# Can be assumed that the campaign is not disabled now
 
 			# Page Vars
 			page_vars = { 'page_title': 'Upload Report'}
@@ -88,12 +104,6 @@ def upload_recipients(request, param_username, param_campaign_pk):
 			valid_recip = True
 			valid_email_count = 0
 			invalid_email_count = 0
-
-			# Capturing Campaign that will be used with the recipient entries
-			try:
-				campaign = Campaign.objects.get(pk=param_campaign_pk, user=request.user)
-			except Campaign.DoesNotExist:
-				raise Http404
 
 			form = ContactUploadForm(request.POST, request.FILES)
 
@@ -183,21 +193,31 @@ def add_recipient(request, param_username, param_campaign_pk):
 
 		# User is uploading from the form
 		elif request.method == "POST":
-			completed_form = RecipientForm(request.POST)
 
-			if completed_form.is_valid():
-				recipient = completed_form.save(commit=False)
+			# Check to see if the campaign is disabled
+			if page_vars['campaign'].active:
+				completed_form = RecipientForm(request.POST)
 
-				recipient.campaign = page_vars['campaign']
+				if completed_form.is_valid():
+					recipient = completed_form.save(commit=False)
 
-				# Send out invitation!!!
-				os.system("python manage.py deploy_invitations {0} &".format( recipient.campaign.pk ))
+					recipient.campaign = page_vars['campaign']
 
-				recipient.save()
-				return HttpResponseRedirect('/' + request.user.username + '/campaign-' + str(page_vars['campaign'].pk) + '/recipients')
+					# Send out invitation!!!
+					os.system("python manage.py deploy_invitations {0} &".format( recipient.campaign.pk ))
+
+					recipient.save()
+					return HttpResponseRedirect('/' + request.user.username + '/campaign-' + str(page_vars['campaign'].pk) + '/recipients')
+				else:
+					generate_form_errors(request, completed_form)
+					page_vars['form'] = completed_form
+					return render(request, 'recipient/add.html', page_vars)
+
 			else:
-				generate_form_errors(request, completed_form)
-				page_vars['form'] = completed_form
-				return render(request, 'recipient/add.html', page_vars)
+
+				# Campaign is deactivated, generate message.
+				messages.add_message(request, messages.SUCCESS, 'Campaign disabled by MailPrime Administrator')
+				return HttpResponseRedirect('/{0}/campaign-{1}/recipients/add'.format(request.user.username, page_vars['campaign'].pk))
+
 	else:
 		raise Http404
